updated testing
